Streaming Chat
We now support streaming! The backend can send up the right headers and content values to have the response streamed back to the user. This requires modifying nginx.conf to work or else nginx will buffer between the backend and the frontend.

The frontend has been refactored to use fetch for the chat request because axios does not seem to support streaming. Processes only `data:` messages and account for messages split across reads. Note: When you await fetch, the promise is fulfilled when the browser gets the headers back.