Tool Execution
Changed our approach to tools. Instead of adding descriptions of tools to the prompt, we're using the API provided mechanisms.

Updated options for Chat so tool use can be automatic or manual in soc. ChatStream will be used for direct user interaction and should always allow the UI to determine if a tool gets executed or not.

Cleaned up some of the documentation contents for our only tool.