Check if non-community repo rules exist during sync
When syncing ElastAlert detections, if a non-community rule already exists in the DB: do nothing. This was previously creating duplicates of non-community repo rules, and later colliding with existing rules after we used determinism to stop creating duplicates, but now it'll cleanly avoid creating the detection and log that it was already present.

Added tests for new cases.