Writing Chat History
We need to use a different structure for storing a message than we do for transferring one. Models have been updated to account for this.

Parsing and saving a chat message after the client has disconnected needed refactoring. The context was being canceled by the disconnected client so the save wouldn't go through.

Improved validation.