Tags and DeletedAt on Model, Delete Endpoint
Elasticsearch storage models have been updated to allow tagging, deleting, and the storage of tool usage parameters.

A new endpoint has been added that allows for the deleting of chat sessions. Because sessions are ethereal, special settings are specified on the first message of the session. Calls for previous conversations will not return deleted sessions.