ES-Backed Message History
Refactored UI to stop building message history and sending it up with chat and tool requests.

Updated models to match what the AI Gateway and AI API expect. Removed a lot of unnecessary complexity.

Updated handlers to retrieve history by sessionId rather than expecting it to be passing in by the request. Updated tests.

Updated AssistantManager interface for non-streaming chat to return all messages that get created (i.e. automatic tool use and tool result messages).

Solved ToolResult. Ultimately we're sending it back to the AI as a user message. This appears to be the only way to get the API to not error and response to a tool result. These messages are now tagged with `tool_result` so the UI can easily identify them.