get_playbook_questions tool, more reliable tool results
get_playbook_questions mostly implemented. It fetches the questions and simplifies them down so the investigation can continue. I'll be working to expand this to include the answers to the questions but checking in now so the frontend has something to work with.

The assistanthandler now works harder to ensure that a success or error message is always added to the session.