get_playbook_questions now serves results
Improved tool so it fetches the answers to the questions.

Updated Investigation prompt to adapt to tool changes.

Added and updated tests.

Eventually, the current back-and-forth playbook conversion (send the playbook to the UI, UI substitutes variables, send the questions to the backend, backend converts and runs queries) that takes place in Alerts will be refactored to use this new backend-only code. But I'm going to leave them separate for a bit as we gain confidence in this new implementation.