Split Models for Chat vs Session
Originally we were putting metadata for a session on the first message of that session. Now we have dedicated session objects in a new ES index.

A few frontend changes were made to adapt to these changes. Updated tests too.

Anybody running the assistant before this commit should remove the so-assistant index.