Health check during Balance
Do a health check before requesting the balance. Return the health status in the same response. Includes new config value for the timeout that defaults to 3 seconds. Update tests.

Also changed Balance's Credits to be an int64. Confirmed with Mike that it's never fractional and can grow quite large. Refactored tests.

Regenerated mocks.