Aggregate Support
Remove unused functions from JS. Cleanup unused vars.

Consolidate aggregate and non-aggregate payloads reducing code duplication.

TODO: Currently looking into the necessity of removing the * from groupby clauses in playbook.go. This wasn't necessary in .190 and isn't explicitly mentioned in the playbook source. When it's present in queries that group by e.g. destination.ip, it specifies that any missing values should use the value of "__missing__" which isn't a valid IP so the query fails.

But for now it works.