Reorchestration, UI Work
Remove endpoint in favor of UI orchestrated message being sent on behalf of the user. This message will be tagged and the tag used to facilitate the compression.

Show a divider above the latest context_compression message.

Update how Context Length is calculated to reset when a context_compression message is seen.

Moved button from top toolbar to next to the Context Length label.

Updated tests.