Allow Compression when at Context Limit
According to an AWS blog post (https://aws.amazon.com/blogs/security/context-window-overflow-breaking-the-barrier/) context begins falling off the beginning when the content of a session grows larger than the context. Normally we prevent a user from sending messages when in this situation but compressing the context could keep the conversation going in a useful fashion.

Update tests.