Toggle Session Shared Tag
This takes the first steps towards conversations being accessible by other users. The original owner of a session can mark a session as shared by tagging it. In later commits, we'll use this tag when determining if a user can access a session that isn't theirs.

The UI is ugly but functional.

Tests are needed.