Stable Working State
Ripped out a lot of the GetSessions logic that checks perms ahead of time. Now it executes the search and filters the result set. No longer need to indicate before hand if the search is looking for authored sessions or not.

Fix existing tests so they pass. Will look into adding more after some cleanup.