Silent Parroting
Because Amazon Bedrock gives us messages that it won't take, specifically a message from the AI containing ONLY a tool request and no text, we speak on behalf of the AI when we correct it. This induces parroting. Until parroting can be properly fixed, I'm giving it a "non-whitespace whitespace" string consisting of a non-breaking space. It'll render how we want in the frontend.