Attach Session To Case
Add an escalate triangle to the assistant page, only show it when a session has been loaded. Clicking it offers the same mruCases list we use in Alerts including creating a new case.

Case has been updated to show Assistant Sessions slightly differently from other attachments such as hiding hashes.

To do: Still need to modify UpdateSession endpoint to not allow the removal of the "shared" tag if the session is attached to a case. Tests.